+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Updated user agent string formats to allow viewing version distributions in the new PowerSync dashboard.
 - Sync options: `newClientImplementation` is now the default.
 - Make `androidx.sqlite:sqlite-bundled` an API dependency of `:core` to avoid toolchain warnings.
+- On Apple platforms, use a websocket protocol as a workaround to clients not supporting backpressure in HTTP response
+  streams.
 
 ## 1.8.1